1.portion change to quantity&unit 2.use DateUtil to format date 3.TODO: modify code of addIngredient to under realm writing but need better structure

+++ DietLens/Services/ProfileDataManager.swift
@@ -23,8 +23,8 @@ class ProfileDataManager {
         userProfile.occupation = jsonObj["occupation"].stringValue
         userProfile.nationality = jsonObj["nationality"].stringValue
         userProfile.marital_status = jsonObj["marital_status"].stringValue
-        userProfile.height = Double(jsonObj["height"].intValue)
-        userProfile.weight = Double(jsonObj["weight"].intValue)
+        userProfile.height = jsonObj["height"].doubleValue
+        userProfile.weight = jsonObj["weight"].doubleValue
         return userProfile
     }
 }
